NewCodeStructureToFollow
