+++ src/main/java/logic/commands/Command.java
@@ -1,4 +1,4 @@
-package commands;
+package main.java.logic.commands;
 
 public interface Command {
