+++ src/main/java/logic/commands/close/CloseADoor.java
@@ -1,19 +1,19 @@
-package commands.open;
+package main.java.logic.commands.close;
 
-import commands.Command;
-import rooms.Room;
+import main.java.logic.commands.Command;
+import main.java.model.rooms.Room;
 
-public class OpenADoor implements Command{
+public class CloseADoor implements Command{
     protected Room room;
 
-    public OpenADoor(Room room){
+    public CloseADoor(Room room){
         this.room = room;
     }
 
     @Override
     public Boolean execute(){
         // if(room.getHasDoors()){
-            // room.setDoorOpen(true);
+            // room.setDoorOpen(false);
         // } else {
             // potentially send a GUI error
             // System.out.println("No doors to open in " + room);
