+++ src/main/java/logic/modules/Module.java
@@ -1,0 +1,9 @@
+package main.java.logic.modules;
+
+import main.java.logic.observerPattern.Observable;
+import main.java.logic.observerPattern.Observer;
+
+public abstract class Module implements Observer{
+    @Override
+    public abstract void update(Observable o);
+}
