+++ src/main/java/logic/observerPattern/Observer.java
@@ -1,4 +1,4 @@
-package observerPattern;
+package main.java.logic.observerPattern;
 
 public interface Observer {
     void update(Observable o);
