--- src/modules/Module.java
@@ -1,9 +1,0 @@
-package modules;
-
-import observerPattern.Observer;
-import observerPattern.Observable;
-
-public abstract class Module implements Observer{
-    @Override
-    public abstract void update(Observable o);
-}
